Move SvgEmptyState to private dir and use CSS module for styles
Move the component from src/components/svg-empty-state/ to
src/charts/private/svg-empty-state/ alongside other internal chart
components. Replace inline style object with a SCSS module and use
clsx for conditional class composition.
