+++ projects/js-packages/charts/src/charts/private/svg-empty-state/index.ts
