+++ projects/js-packages/charts/src/charts/private/svg-empty-state/svg-empty-state.tsx
@@ -1,5 +1,7 @@
-import { isSafari } from '../../utils';
-import type { FC, ReactNode, CSSProperties } from 'react';
+import clsx from 'clsx';
+import { isSafari } from '../../../utils';
+import styles from './svg-empty-state.module.scss';
+import type { FC, ReactNode } from 'react';
 
 interface SvgEmptyStateProps {
 	/** X coordinate of the center point */
@@ -41,24 +43,17 @@ export const SvgEmptyState: FC< SvgEmptyStateProps > = ( {
 	fill = '#ccc',
 	children,
 } ) => {
-	const textStyles: CSSProperties = {
-		display: 'flex',
-		alignItems: 'center',
-		justifyContent: 'center',
-		textAlign: 'center',
-		width: '100%',
-		height: '100%',
-		color: fill,
-		fontSize: '14px',
-		fontFamily: '-apple-system,BlinkMacSystemFont,Roboto,Helvetica Neue,sans-serif',
-		padding: '16px',
-		boxSizing: 'border-box',
-		...( isSafari() ? { position: 'fixed' as const } : {} ),
-	};
-
 	return (
 		<foreignObject x={ x - width / 2 } y={ y - height / 2 } width={ width } height={ height }>
-			<div style={ textStyles }>{ children }</div>
+			<div
+				className={ clsx(
+					styles[ 'svg-empty-state' ],
+					isSafari() && styles[ 'svg-empty-state--safari' ]
+				) }
+				style={ { color: fill } }
+			>
+				{ children }
+			</div>
 		</foreignObject>
 	);
 };
